+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 - Input: fix PointerEvent detection on Firefox for Android mobile (thanks @kutyamutya)
 - Renderer: fix a regression with strokeLine in the CanvasRenderer
+- Geometry: fix the RoundRect clone method
 
 ## [10.10.0] (melonJS 2) - _2022-06-07_